Update Datatable.jsx
When the code was const [list, setList] = useState(); it created an error as length of initial state i.e. undefined was accessed.

const [list, setList] = useState([]);  initializes the list state variable with an empty array and provides a function setList to update the state. This solves the issues, as now length can be accessed.